+++ extensions-support/langchain4j/deployment/src/main/java/org/apache/camel/quarkus/component/support/langchain4j/deployment/SupportLangchain4jProcessor.java
@@ -18,12 +18,13 @@ package org.apache.camel.quarkus.component.support.langchain4j.deployment;
 
 import java.util.List;
 
-import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.chat.ChatModel;
 import io.quarkiverse.langchain4j.deployment.items.SelectedChatModelProviderBuildItem;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.annotations.ExecutionTime;
 import io.quarkus.deployment.annotations.Record;
+import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
 import org.apache.camel.quarkus.component.support.langchain4j.Langchain4jRecorder;
 import org.apache.camel.quarkus.core.deployment.spi.CamelBeanQualifierResolverBuildItem;
 
@@ -38,9 +39,28 @@ class SupportLangchain4jProcessor {
         // Enable ChatLanguageModel instances to be resolved by name from the Camel registry
         for (SelectedChatModelProviderBuildItem chatModel : chatModels) {
             beanQualifierResolver.produce(
-                    new CamelBeanQualifierResolverBuildItem(ChatLanguageModel.class,
+                    new CamelBeanQualifierResolverBuildItem(ChatModel.class,
                             chatModel.getConfigName(),
                             recorder.chatModelBeanQualifierResolver(chatModel.getConfigName())));
         }
+    }
+
+    // TODO: Remove this: https://github.com/apache/camel-quarkus/issues/7440
+    @BuildStep
+    ReflectiveClassBuildItem registerForReflection() {
+        return ReflectiveClassBuildItem.builder(
+                "dev.langchain4j.model.ollama.FormatSerializer",
+                "dev.langchain4j.model.ollama.Function",
+                "dev.langchain4j.model.ollama.FunctionCall",
+                "dev.langchain4j.model.ollama.Message",
+                "dev.langchain4j.model.ollama.OllamaChatRequest",
+                "dev.langchain4j.model.ollama.OllamaChatResponse",
+                "dev.langchain4j.model.ollama.Options",
+                "dev.langchain4j.model.ollama.Parameters",
+                "dev.langchain4j.model.ollama.Role",
+                "dev.langchain4j.model.ollama.Tool",
+                "dev.langchain4j.model.ollama.ToolCall")
+                .methods(true)
+                .build();
     }
 }