+++ extensions/openapi-java/deployment/src/main/java/org/apache/camel/quarkus/component/openapi/java/deployment/OpenApiJavaProcessor.java
@@ -94,7 +94,9 @@ class OpenApiJavaProcessor {
             Capabilities capabilities) throws Exception {
 
         if (capabilities.isPresent(Capability.SMALLRYE_OPENAPI)) {
-            RoutesConfigurer configurer = new RoutesConfigurer();
+            final CamelContext ctx = CamelSupport.newBuildTimeCamelContext(true);
+
+            RoutesConfigurer configurer = new RoutesConfigurer(ctx);
             List<RoutesBuilder> routes = new ArrayList<>();
             configurer.setRoutesBuilders(routes);
             configurer.setRoutesCollector(new DefaultRoutesCollector());
@@ -103,7 +105,6 @@ class OpenApiJavaProcessor {
             configurer.setRoutesExcludePattern(
                     CamelMainHelper.routesExcludePattern().collect(Collectors.joining(",")));
 
-            final CamelContext ctx = CamelSupport.newBuildTimeCamelContext(true);
             if (!routesBuilderClasses.isEmpty()) {
                 final ClassLoader loader = Thread.currentThread().getContextClassLoader();
                 if (!(loader instanceof QuarkusClassLoader)) {