+++ integration-test-groups/azure/azure-key-vault/src/test/java/org/apache/camel/quarkus/component/azure/key/vault/it/AzureKeyVaultTest.java
@@ -70,12 +70,4 @@ class AzureKeyVaultTest extends AbstractAzureKeyVaultTest {
             }
         }
     }
-
-    @Test
-    void dynamicExceptionInstantiation() {
-        RestAssured.get("/azure-key-vault/exception/cache")
-                .then()
-                .statusCode(200)
-                .body(is("true"));
-    }
 }